Co-locate list resources under internal/resource/{profile,document}
Move the per-resource list adapters and result-assembly translators
from internal/listresource/{profile,document}/ into the matching
managed-resource packages as list.go + list_test.go, following the
repo convention of one package per Terraform resource type with
operation-named files.

The shared paginate domain helper stays at internal/listresource/
since it is genuinely cross-cutting (used by both packages and
indifferent to which resource type it walks). Function names lose
their now-redundant Profile/Document qualifiers since they are
already namespaced by package — buildProfileListResult is now
profile.buildListResult, etc.
