+++ internal/resource/document/list.go
@@ -1,4 +1,4 @@
-package documentlist
+package document
 
 import (
 	"context"
@@ -7,8 +7,9 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-framework/diag"
 	"github.com/hashicorp/terraform-plugin-framework/list"
-	"github.com/hashicorp/terraform-plugin-framework/list/schema"
+	listschema "github.com/hashicorp/terraform-plugin-framework/list/schema"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
+	"github.com/hashicorp/terraform-plugin-framework/types"
 
 	"github.com/dmalch/terraform-provider-genealogy/internal/config"
 	"github.com/dmalch/terraform-provider-genealogy/internal/geni"
@@ -31,7 +32,7 @@ func (r *listResource) Metadata(_ context.Context, req resource.MetadataRequest,
 }
 
 func (r *listResource) ListResourceConfigSchema(_ context.Context, _ list.ListResourceSchemaRequest, resp *list.ListResourceSchemaResponse) {
-	resp.Schema = schema.Schema{}
+	resp.Schema = listschema.Schema{}
 }
 
 func (r *listResource) Configure(_ context.Context, req resource.ConfigureRequest, resp *resource.ConfigureResponse) {
@@ -68,6 +69,62 @@ func streamUploadedDocuments(ctx context.Context, c *geni.Client, req list.ListR
 			}}
 		},
 		func(d geni.DocumentResponse) (list.ListResult, bool) {
-			return buildDocumentListResult(ctx, &d, req)
+			return buildListResult(ctx, &d, req)
 		})
+}
+
+// displayNameFor produces a human-readable label for a document in query
+// output. The Title is the obvious choice; we fall back to the bare ID when
+// it is absent so the result remains visually identifiable.
+func displayNameFor(d *geni.DocumentResponse) string {
+	if d.Title == "" {
+		return d.Id
+	}
+	return fmt.Sprintf("%s (%s)", d.Title, d.Id)
+}
+
+// buildListResult turns one API response into a list.ListResult whose Identity
+// carries the document ID under the managed resource's identity schema. When
+// req.IncludeResource is true the Resource field is populated via ValueFrom —
+// the same translator used by Read — so list output round-trips through
+// `import { identity = ... }`.
+func buildListResult(
+	ctx context.Context,
+	resp *geni.DocumentResponse,
+	req list.ListRequest,
+) (list.ListResult, bool) {
+	result := req.NewListResult(ctx)
+
+	identity := ResourceIdentityModel{
+		ID: types.StringValue(resp.Id),
+	}
+	diags := result.Identity.Set(ctx, identity)
+	result.Diagnostics.Append(diags...)
+	if result.Diagnostics.HasError() {
+		return result, false
+	}
+
+	result.DisplayName = displayNameFor(resp)
+
+	if req.IncludeResource {
+		// Seed every collection field with a typed null so ValueFrom can
+		// overwrite the fields the API does return without leaving any
+		// untouched attribute carrying a type-less zero value (see the
+		// equivalent comment in the profile package for the failure mode).
+		model := NewEmptyResourceModel()
+
+		diags = ValueFrom(ctx, resp, &model)
+		result.Diagnostics.Append(diags...)
+		if result.Diagnostics.HasError() {
+			return result, false
+		}
+
+		diags = result.Resource.Set(ctx, model)
+		result.Diagnostics.Append(diags...)
+		if result.Diagnostics.HasError() {
+			return result, false
+		}
+	}
+
+	return result, true
 }
