+++ internal/resource/document/list_test.go
@@ -1,4 +1,4 @@
-package documentlist
+package document
 
 import (
 	"testing"
@@ -8,17 +8,14 @@ import (
 	. "github.com/onsi/gomega"
 
 	"github.com/dmalch/terraform-provider-genealogy/internal/geni"
-	documentresource "github.com/dmalch/terraform-provider-genealogy/internal/resource/document"
 )
 
-func ptr[T any](v T) *T { return &v }
-
-// listRequestForDocument builds a list.ListRequest carrying the live
-// managed-resource schemas. The caller must have already registered gomega
-// for the current test via RegisterTestingT.
-func listRequestForDocument(t *testing.T, includeResource bool) list.ListRequest {
+// listRequest builds a list.ListRequest carrying the live managed-resource
+// schemas. The caller must have already registered gomega for the current
+// test via RegisterTestingT.
+func listRequest(t *testing.T, includeResource bool) list.ListRequest {
 	t.Helper()
-	r := documentresource.NewResource()
+	r := NewResource()
 
 	var schemaResp resource.SchemaResponse
 	r.Schema(t.Context(), resource.SchemaRequest{}, &schemaResp)
@@ -52,50 +49,50 @@ func TestDisplayNameFor(t *testing.T) {
 	})
 }
 
-func TestBuildDocumentListResult(t *testing.T) {
+func TestBuildListResult(t *testing.T) {
 	t.Run("Populates Identity with the document ID", func(t *testing.T) {
 		RegisterTestingT(t)
-		req := listRequestForDocument(t, false)
+		req := listRequest(t, false)
 		givenResponse := &geni.DocumentResponse{Id: "document-42", Title: "Test Doc"}
 
-		result, ok := buildDocumentListResult(t.Context(), givenResponse, req)
+		result, ok := buildListResult(t.Context(), givenResponse, req)
 
 		Expect(ok).To(BeTrue())
 		Expect(result.Diagnostics.HasError()).To(BeFalse())
 		Expect(result.Identity).NotTo(BeNil())
 
-		var identity documentresource.ResourceIdentityModel
+		var identity ResourceIdentityModel
 		Expect(result.Identity.Get(t.Context(), &identity).HasError()).To(BeFalse())
 		Expect(identity.ID.ValueString()).To(Equal("document-42"))
 	})
 
 	t.Run("Sets a human-readable DisplayName", func(t *testing.T) {
 		RegisterTestingT(t)
-		req := listRequestForDocument(t, false)
+		req := listRequest(t, false)
 		givenResponse := &geni.DocumentResponse{Id: "document-42", Title: "Test Doc"}
 
-		result, ok := buildDocumentListResult(t.Context(), givenResponse, req)
+		result, ok := buildListResult(t.Context(), givenResponse, req)
 
 		Expect(ok).To(BeTrue())
 		Expect(result.DisplayName).To(Equal("Test Doc (document-42)"))
 	})
 
-	t.Run("Populates Resource via document.ValueFrom when IncludeResource is true", func(t *testing.T) {
+	t.Run("Populates Resource via ValueFrom when IncludeResource is true", func(t *testing.T) {
 		RegisterTestingT(t)
-		req := listRequestForDocument(t, true)
+		req := listRequest(t, true)
 		givenResponse := &geni.DocumentResponse{
 			Id:          "document-43",
 			Title:       "Birth Certificate",
 			ContentType: ptr("image/png"),
 		}
 
-		result, ok := buildDocumentListResult(t.Context(), givenResponse, req)
+		result, ok := buildListResult(t.Context(), givenResponse, req)
 
 		Expect(ok).To(BeTrue())
 		Expect(result.Diagnostics.HasError()).To(BeFalse())
 		Expect(result.Resource).NotTo(BeNil())
 
-		var model documentresource.ResourceModel
+		var model ResourceModel
 		Expect(result.Resource.Get(t.Context(), &model).HasError()).To(BeFalse())
 		Expect(model.ID.ValueString()).To(Equal("document-43"))
 		Expect(model.Title.ValueString()).To(Equal("Birth Certificate"))
